refactor companyStep and IdentityForm to use new input
