+++ src/pages/settings/Payments/AddDebitCardPage.js
@@ -19,12 +19,11 @@ import KeyboardAvoidingView from '../../../components/KeyboardAvoidingView';
 import * as ValidationUtils from '../../../libs/ValidationUtils';
 import CheckboxWithLabel from '../../../components/CheckboxWithLabel';
 import ExpensiTextInput from '../../../components/ExpensiTextInput';
-import AddressSearchInput from '../../AddressSearchInput';
+import AddressSearchInput from '../../../components/AddressSearchInput';
 import CONST from '../../../CONST';
 import FormAlertWithSubmitButton from '../../../components/FormAlertWithSubmitButton';
 import ONYXKEYS from '../../../ONYXKEYS';
 import compose from '../../../libs/compose';
-import ROUTES from '../../../ROUTES';
 
 const propTypes = {
     addDebitCardForm: PropTypes.shape({
@@ -227,9 +226,7 @@ class DebitCardPage extends Component {
                                 label={this.props.translate('addDebitCardPage.billingAddress')}
                                 containerStyles={[styles.mt4]}
                                 onChange={(fieldName, value) => this.clearErrorAndSetValue(fieldName, value)}
-                                value={this.state.addressStreet}
                                 errorText={this.getErrorText('addressStreet')}
-                                onFocus={() => Navigation.navigate(ROUTES.ADDRESS_SEARCH)}
                             />
                             <CheckboxWithLabel
                                 isChecked={this.state.acceptedTerms}
